Separate RestrictedWidget
