+++ client/app/components/dashboards/dashboard-widget/index.js
@@ -1,2 +1,3 @@
 export { default as VisualizationWidget } from './VisualizationWidget';
 export { default as TextboxWidget } from './TextboxWidget';
+export { default as RestrictedWidget } from './RestrictedWidget';
